% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_reg_lbfgs}
\alias{lin_reg_lbfgs}
\title{Linear Regression with L-BFGS}
\usage{
lin_reg_lbfgs(X, y)
}
\arguments{
\item{X}{A \code{matrix} that is the Design Matrix for the regression problem.}

\item{y}{A \code{vec} containing the response values.}
}
\description{
Optimizes Linear Regression's Residual Sum of Squares.
}
\details{
The objective function is defined as:

\deqn{f(\theta) = (y - \theta' X)^2}{f(theta) = (y - theta' X)^2}

The gradient is defined as:

\deqn{\dfrac{\partial\text{RSS}}{\partial \beta} = -2\mathbf{X}^{T}\left(\mathbf{y}-\mathbf{X}\beta\right)}{dRSS/dBeta = -2 X' (y - X theta)}
}
\examples{
# Number of Points
n = 1000

# Select beta parameters
beta = c(-2, 1.5, 3, 8.2, 6.6)

# Number of Predictors (including intercept)
p = length(beta)

# Generate predictors from a normal distribution
X_i = matrix(rnorm(n), ncol = p - 1)

# Add an intercept
X = cbind(1, X_i)

# Generate y values
y = X\%*\%beta + rnorm(n / (p - 1))

# Run optimization with lbfgs
theta_hat = lin_reg_lbfgs(X, y)

# Verify parameters were recovered
cbind(actual = beta, estimated = theta_hat)
}
