% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppEnsmallen-package.R
\docType{package}
\name{RcppEnsmallen-package}
\alias{RcppEnsmallen}
\alias{RcppEnsmallen-package}
\title{RcppEnsmallen: Header-only C++ Mathematical Optimization Library for 'Armadillo'}
\description{
'Ensmallen' is a templated C++ mathematical optimization library 
 (by the MLPACK team) that provides a simple set of abstractions for writing an
 objective function to optimize. Provided within are various standard and
 cutting-edge optimizers that include full-batch gradient descent techniques, 
 small-batch techniques, gradient-free optimizers, and constrained optimization.
 The 'RcppEnsmallen' package includes the header files from the 'Ensmallen' library
 and pairs the appropriate header files from 'armadillo' through the 
 'RcppArmadillo' package. Therefore, users do not need to install 'Ensmallen' nor
 'Armadillo' to use 'RcppEnsmallen'. Note that 'Ensmallen' is licensed under 
 3-Clause BSD, 'Armadillo' starting from 7.800.0 is licensed under Apache License 2,
 'RcppArmadillo' (the 'Rcpp' bindings/bridge to 'Armadillo') is licensed under 
 the GNU GPL version 2 or later. Thus, 'RcppEnsmallen' is also licensed under
 similar terms. Note that 'Ensmallen' requires a compiler that supports 
 'C++11' and 'Armadillo' 6.500 or later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/RcppEnsmallen}
  \item \url{https://github.com/mlpack/ensmallen}
  \item \url{http://ensmallen.org/}
  \item Report bugs at \url{https://github.com/coatless/RcppEnsmallen/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Authors:
\itemize{
  \item Dirk Eddelbuettel \email{edd@debian.org} (0000-0001-6419-907X) [copyright holder]
  \item Conrad Sanderson \email{conradsand@gmail.com} [copyright holder]
  \item Ryan Curtin \email{ryan@ratml.org} [copyright holder]
  \item Ensmallen and MLPACK Development Team [copyright holder]
}

}
